failsafer score calc and tests

+++ jahmm/test/utils/TestParameters.java
@@ -9,7 +9,7 @@ import java.util.logging.Logger;
 public final class TestParameters {
 
     public static final double EPSILON = 1e-06d;
-    public static final double NUMBER_OF_TESTS = 0x1000;
+    public static final int NUMBER_OF_TESTS = 0x1000;
     private static final Logger LOG = Logger.getLogger(TestParameters.class.getName());
 
     private TestParameters() {
